[OPIK-601] [UX improvements] Rename experiment tabs
