+++ apps/opik-frontend/src/components/pages/CompareExperimentsPage/CompareExperimentsPanel/DataTab/DataTab.tsx
@@ -1,19 +1,50 @@
 import React from "react";
-import { DatasetItem } from "@/types/datasets";
-import SyntaxHighlighter from "@/components/shared/SyntaxHighlighter/SyntaxHighlighter";
-import NoData from "@/components/shared/NoData/NoData";
+import {
+  ResizableHandle,
+  ResizablePanel,
+  ResizablePanelGroup,
+} from "@/components/ui/resizable";
+
+import { DatasetItem, ExperimentItem } from "@/types/datasets";
+import CompareExperimentsViewer from "@/components/pages/CompareExperimentsPage/CompareExperimentsPanel/CompareExperimentsViewer";
+import { OnChangeFn } from "@/types/shared";
+import ExperimentDataset from "@/components/pages/CompareExperimentsPage/CompareExperimentsPanel/DataTab/ExperimentDataset";
 
 interface DataTabProps {
   data: DatasetItem["data"] | undefined;
+  experimentItems: ExperimentItem[];
+  openTrace: OnChangeFn<string>;
 }
 
-const DataTab = ({ data }: DataTabProps) => {
-  return data ? (
-    <div className="pb-8">
-      <SyntaxHighlighter data={data} />
-    </div>
-  ) : (
-    <NoData />
+const DataTab = ({ data, experimentItems, openTrace }: DataTabProps) => {
+  const renderExperimentsSection = () => {
+    return experimentItems.map((experimentItem, idx) => (
+      <React.Fragment key={experimentItem.id}>
+        <ResizablePanel className="min-w-72" style={{ overflow: "unset" }}>
+          <CompareExperimentsViewer
+            experimentItem={experimentItem}
+            openTrace={openTrace}
+          />
+        </ResizablePanel>
+
+        {idx !== experimentItems.length - 1 ? <ResizableHandle /> : null}
+      </React.Fragment>
+    ));
+  };
+
+  return (
+    <ResizablePanelGroup
+      direction="horizontal"
+      autoSaveId="compare-vetical-sidebar"
+      style={{ height: "unset", overflow: "unset" }}
+      className="min-h-full"
+    >
+      <ResizablePanel defaultSize={30} className="min-w-72">
+        <ExperimentDataset data={data} />
+      </ResizablePanel>
+      <ResizableHandle />
+      {renderExperimentsSection()}
+    </ResizablePanelGroup>
   );
 };
